Fix UE 5.5 errors 1. Fix Incomplete Type Error with std::unique_ptr and Forward Declaration 2. Fix format specifier mismatch in UE_LOG causing static_assert failure
